+++ Models/SyncFolder.swift
@@ -15,8 +15,6 @@ struct SyncFolder: Codable, Identifiable, Equatable, Hashable {
     var isEnabled: Bool
     var excludePatterns: [String]
     var lastSyncDate: Date?
-    var syncSchedule: SyncSchedule?
-    var fileFilters: FileFilters?
     var symlinkMode: Bool
     var symlinkState: SymlinkState
     var symlinkProtected: Bool
@@ -29,8 +27,6 @@ struct SyncFolder: Codable, Identifiable, Equatable, Hashable {
          isEnabled: Bool = true,
          excludePatterns: [String] = SyncFolder.defaultExcludePatterns,
          lastSyncDate: Date? = nil,
-         syncSchedule: SyncSchedule? = nil,
-         fileFilters: FileFilters? = nil,
          symlinkMode: Bool = true,
          symlinkState: SymlinkState = .local,
          symlinkProtected: Bool = false) {
@@ -42,8 +38,6 @@ struct SyncFolder: Codable, Identifiable, Equatable, Hashable {
         self.isEnabled = isEnabled
         self.excludePatterns = excludePatterns
         self.lastSyncDate = lastSyncDate
-        self.syncSchedule = syncSchedule
-        self.fileFilters = fileFilters
         self.symlinkMode = symlinkMode
         self.symlinkState = symlinkState
         self.symlinkProtected = symlinkProtected
@@ -59,8 +53,6 @@ struct SyncFolder: Codable, Identifiable, Equatable, Hashable {
         isEnabled = try container.decode(Bool.self, forKey: .isEnabled)
         excludePatterns = try container.decode([String].self, forKey: .excludePatterns)
         lastSyncDate = try container.decodeIfPresent(Date.self, forKey: .lastSyncDate)
-        syncSchedule = try container.decodeIfPresent(SyncSchedule.self, forKey: .syncSchedule)
-        fileFilters = try container.decodeIfPresent(FileFilters.self, forKey: .fileFilters)
         symlinkMode = try container.decodeIfPresent(Bool.self, forKey: .symlinkMode) ?? true
         symlinkState = try container.decodeIfPresent(SymlinkState.self, forKey: .symlinkState) ?? .local
         symlinkProtected = try container.decodeIfPresent(Bool.self, forKey: .symlinkProtected) ?? false
@@ -109,41 +101,4 @@ struct SyncFolder: Codable, Identifiable, Equatable, Hashable {
             nasPath: "/Volumes/home/Desktop"
         )
     ]
-}
-
-struct SyncSchedule: Codable, Equatable, Hashable {
-    var startHour: Int
-    var endHour: Int
-    var daysOfWeek: [Int]
-    var onlyOnWiFi: Bool
-    var wifiSSIDs: [String]
-    var onlyOnACPower: Bool
-    
-    init(startHour: Int = 0,
-         endHour: Int = 24,
-         daysOfWeek: [Int] = [0, 1, 2, 3, 4, 5, 6],
-         onlyOnWiFi: Bool = false,
-         wifiSSIDs: [String] = [],
-         onlyOnACPower: Bool = false) {
-        self.startHour = startHour
-        self.endHour = endHour
-        self.daysOfWeek = daysOfWeek
-        self.onlyOnWiFi = onlyOnWiFi
-        self.wifiSSIDs = wifiSSIDs
-        self.onlyOnACPower = onlyOnACPower
-    }
-}
-
-struct FileFilters: Codable, Equatable, Hashable {
-    var allowedExtensions: [String]
-    var maxFileSizeGB: Double?
-    var minFileSizeKB: Double?
-    
-    init(allowedExtensions: [String] = [],
-         maxFileSizeGB: Double? = nil,
-         minFileSizeKB: Double? = nil) {
-        self.allowedExtensions = allowedExtensions
-        self.maxFileSizeGB = maxFileSizeGB
-        self.minFileSizeKB = minFileSizeKB
-    }
 }